Add migration

+++ Machete.Domain/Entities/Worker.cs
@@ -200,11 +200,10 @@ namespace Machete.Domain
         //
         [LocalizedDisplayName("skill3", NameResourceType = typeof(Resources.Worker))]
         public int? skill3 { get; set; }
-        // TODO: Add worker rating capability
-        /*
+
         [LocalizedDisplayName("workerRating", NameResourceType = typeof(Resources.Worker))]
-        public int? workerRating { get; set; }
-        */
+        public float? workerRating { get; set; }
+
         //
         public bool isActive 
         {
